Add warnings/errors for bad download

+++ YOCdownload.py
@@ -4,27 +4,35 @@
 # Python includes.
 import argparse
 from html.parser import HTMLParser
+import itertools
 import urllib.request
 import urllib.parse
 import os
 from multiprocessing import Pool
+# pip3 install --user python-magic
+# python-magic library: https://github.com/ahupp/python-magic
+import magic
 
 # Variables.
 BASEOCURL = "http://ocremix.org/remix/OCR0"
 
 # Get arguments
 parser = argparse.ArgumentParser(description='Download a numeric range of mixes from ocremix.org.')
-parser.add_argument('ocstart', type=int,
-                    help='the starting oc-remix number')
-parser.add_argument('ocend', type=int,
-                    help='the ending oc-remix number')
+parser.add_argument('ocstart', type=int, help='the starting oc-remix number')
+parser.add_argument('ocend', type=int, help='the ending oc-remix number')
+parser.add_argument("-f", "--forcemirror", help='Force mirror number.', type=int, default=0)
 
 # Save arguments.
 args = parser.parse_args()
 ocstart = args.ocstart
 ocend = args.ocend
 print("Start OCMix: " + format(ocstart) + ", End: " + format(ocend) + ", Total: " + format(ocend - ocstart + 1) + ", Downloading to " + os.getcwd())
+print("Forced Mirror: {0}".format(args.forcemirror))
 input("Press Enter to continue.")
+
+### Global Variables ###
+m = magic.open(magic.MAGIC_NONE)
+m.load()
 
 ### Functions ###
 # Parser documentation: https://docs.python.org/3/library/html.parser.html
@@ -39,7 +47,7 @@ class ImgSrcHTMLParser(HTMLParser):
             self.srcs.append(dict(attrs).get('href'))
 
 
-def ocremix_geturls(mixnumber):
+def ocremix_geturls(mixnumber, force_modulus: int = 0):
     """Store the OCRemix urls, number of mirrors, and modulus."""
     # global ocmp3_urls, ocmirror_number, ocmirror_modulus
     global threadnumber
@@ -63,7 +71,11 @@ def ocremix_geturls(mixnumber):
     ocmirror_number = len(ocmp3_urls)
     # Save to thread number for threadcount.
     threadnumber = len(ocmp3_urls)
-    ocmirror_modulus = mixnumber % ocmirror_number
+    if force_modulus == 0 or force_modulus > threadnumber:
+        ocmirror_modulus = mixnumber % ocmirror_number
+    else:
+        # If forced, set the modulus manually, assuming it is not greater than the threadnumber.
+        ocmirror_modulus = force_modulus - 1
     # Return the modulus url.
     return ocmp3_urls[ocmirror_modulus]
 
@@ -78,21 +90,29 @@ def ocremix_download(url):
     ocfilename = urllib.parse.unquote(os.path.basename(ocfileinfo.path))
 
     # Download the file.
-    print("Downloading mix " + ocfilename + " from mirror " + url)
-    urllib.request.urlretrieve(url, ocfilename)
+    if not os.path.isfile(ocfilename):
+        print("Downloading mix " + ocfilename + " from mirror " + url)
+        urllib.request.urlretrieve(url, ocfilename)
+        # Check mimetype
+        mimetype = m.file(ocfilename)
+        if "Audio" not in mimetype:
+            print("ERROR: Downloaded file {0} is not audio. Deleting.".format(ocfilename))
+            os.remove(ocfilename)
+    else:
+        print("WARNING: File {0} already exists. Skipping.".format(ocfilename))
 
 
 ### Begin Code ###
 
 # Get the info of the first oc mix as a starting reference.
-ocremix_geturls(ocstart)
+ocremix_geturls(ocstart, args.forcemirror)
 # Create threads based on number of mirrors.
 # http://chriskiehl.com/article/parallelism-in-one-line/
 # https://stackoverflow.com/questions/2846653/how-to-use-threading-in-python#2846697
 # Use Pool instead of ThreadPool to really use multiprocessing instead of multiprocessing.dummy.
 pool = Pool(threadnumber)
 # Loop through each mix, ending at ocend. Range is ocend+1 since range function needs to include ocend.
-allurls = pool.map(ocremix_geturls, range(ocstart, ocend + 1))
+allurls = pool.starmap(ocremix_geturls, zip(range(ocstart, ocend + 1), itertools.repeat(args.forcemirror)))
 # Pass the function and array of urls to the pool for downloading.
 results = pool.map(ocremix_download, allurls)
 # Close the queue (no more data will be added to the queue).
